requirements.txt
flask
flask-cors
torch
torchvision
transformers
opencv-python
numpy
pandas
scipy
scikit-image
pillow